updated favicon
Updated favicon for website. Created a pdf-wrapper for balloonProblem pdf file. Pdf will use wrong favicon, and I get a proxy when I have the "proxy" in favicon. Since I am currently not using a server, I'm removing the proxy, so clicking the "Math for Nerds" button does not prompt a proxy error.