Progress on TTT mobile layout
Empty space at the bottom of the screen no longer appears, and the TTT board is centered, with the buttons underneath it, now. However, now, on mobile, when a board has an outcome, the overlay icon indicating the outcome of the board appears in the wrong place. Likely because the overlay is absolutely positioned. Need to fix this.