Mostly functional
You have to really abuse the resizing to get the superTTT page to have some weird formatting, it's pretty good, overall. The Balloon Page needs some work. Need to give it the same 'wide-content' scroll bar treatment. Maybe. Or get the dart throwing animation to not care about the size of the screen.